Move failing Plugin tests into failing_tests sub-directories and ignore tests in those directories
update snapshots after moving tests

The test summary is now:

```
Test Suites: 48 passed, 48 total
Tests:       412 passed, 412 total
Snapshots:   2 passed, 2 total
Time:        12.69s
Ran all test suites.
```
