+++ tests/plugins/failing_tests/channel_header_plug.test.jsx
