+++ tests/plugins/failing_tests/main_menu_action.test.jsx
