+++ tests/plugins/failing_tests/mobile_channel_header_plug.test.jsx
